apps/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
psycopg[binary,pool]==3.2.3
boto3==1.34.160
pydantic-settings==2.4.0
python-dotenv==1.0.1